Add class to old file
